feat: Moved the current default site to statistics and added a new homepage.
