+++ src/Content/UI/Twig/templates/dashboard/empty.html.twig
@@ -1,0 +1,28 @@
+{% extends '@Shared/_layout_vertical.html.twig' %}
+
+{% block title %}{{ 'title.dashboard'|trans }} - {{ parent() }}{% endblock %}
+
+{% block page_header %}
+    <div class="container-xl">
+        <twig:Breadcrumbs :items="[
+            { label: 'title.dashboard' }
+        ]" />
+
+        <div class="page-header d-print-none">
+            <div class="row g-2 align-items-center">
+                <div class="col">
+                    <div class="page-pretitle">{{ 'title.dashboard'|trans }}</div>
+                    <h2 class="page-title">{{ 'dashboard.empty.title'|trans }}</h2>
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock %}
+
+{% block page_content %}
+    <div class="card">
+        <div class="card-body">
+            <p class="text-muted mb-0">{{ 'dashboard.empty.description'|trans }}</p>
+        </div>
+    </div>
+{% endblock %}
