+++ src/Shared/UI/Twig/templates/_embeds/header.html.twig
@@ -17,30 +17,32 @@
                 <div class="d-flex flex-column flex-md-row flex-fill align-items-stretch align-items-md-center">
                     <ul class="navbar-nav">
                         {% block navbar_links %}
-                            <li class="nav-item {{ route starts with 'app_explore' ? 'active' : '' }}">
-                                <a class="nav-link" href="{{ path('app_explore_index') }}">
-                                    <span class="nav-link-title">
-                                        <twig:ux:icon name="tabler:compass" style="height: 1.5em;" />
-                                        {{ 'link.explore'|trans }}
-                                    </span>
-                                </a>
-                            </li>
-                            <li class="nav-item {{ route starts with 'app_stats' ? 'active' : '' }}">
-                                <a class="nav-link" href="{{ path('app_stats_dashboard') }}">
-                                    <span class="nav-link-title">
-                                        <twig:ux:icon name="tabler:graph" style="height: 1.5em;" />
-                                        {{ 'link.statistics'|trans }}
-                                    </span>
-                                </a>
-                            </li>
-                            <li class="nav-item {{ route starts with 'app_import' ? 'active' : '' }}">
-                                <a class="nav-link" href="{{ path('app_import_index') }}">
-                                    <span class="nav-link-title">
-                                        <twig:ux:icon name="tabler:database-import" style="height: 1.5em;" />
-                                        {{ 'link.import'|trans }}
-                                    </span>
-                                </a>
-                            </li>
+                            {% if is_granted('ROLE_USER') %}
+                                <li class="nav-item {{ route starts with 'app_explore' ? 'active' : '' }}">
+                                    <a class="nav-link" href="{{ path('app_explore_index') }}">
+                                        <span class="nav-link-title">
+                                            <twig:ux:icon name="tabler:compass" style="height: 1.5em;" />
+                                            {{ 'link.explore'|trans }}
+                                        </span>
+                                    </a>
+                                </li>
+                                <li class="nav-item {{ route starts with 'app_stats' ? 'active' : '' }}">
+                                    <a class="nav-link" href="{{ path('app_stats_dashboard') }}">
+                                        <span class="nav-link-title">
+                                            <twig:ux:icon name="tabler:graph" style="height: 1.5em;" />
+                                            {{ 'link.statistics'|trans }}
+                                        </span>
+                                    </a>
+                                </li>
+                                <li class="nav-item {{ route starts with 'app_import' ? 'active' : '' }}">
+                                    <a class="nav-link" href="{{ path('app_import_index') }}">
+                                        <span class="nav-link-title">
+                                            <twig:ux:icon name="tabler:database-import" style="height: 1.5em;" />
+                                            {{ 'link.import'|trans }}
+                                        </span>
+                                    </a>
+                                </li>
+                            {% endif %}
                         {% endblock %}
                     </ul>
                 </div>
