+++ src/Statistics/UI/Twig/templates/dashboard/index.html.twig
@@ -10,10 +10,132 @@
     {% endembed %}
 {% endblock %}
 
-{% block page_body %}
+{% block page_header %}
     <div class="container-xl">
         <twig:Breadcrumbs :items="[
-            { label: 'link.statistics' }
+            { label: 'link.statistics' },
+            { label: 'link.stats.dashboard' }
         ]" />
+
+        <div class="page-header d-print-none">
+            <div class="row g-2 align-items-center">
+                <div class="col">
+                    <div class="page-pretitle">
+                        {{ 'link.statistics'|trans }}
+                    </div>
+                    <h2 class="page-title">
+                        {{ 'title.stats.overview'|trans }}
+                    </h2>
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock %}
+
+{% block page_content %}
+    <div class="row row-deck row-cards">
+        <div class="col-sm-6 col-lg-3">
+            <div class="card">
+                <div class="card-body">
+                    <div class="row align-items-center">
+                        <div class="col-auto">
+                            <span class="avatar">
+                                <twig:ux:icon name="tabler:list" style="height: 2em;" />
+                            </span>
+                        </div>
+                        <div class="col">
+                            <div class="font-weight-medium">
+                                {{ allocationCount|number_format(0, ',', '.') }} {{ 'link.allocations'|trans }}
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-6 col-lg-3">
+            <div class="card">
+                <div class="card-body">
+                    <div class="row align-items-center">
+                        <div class="col-auto">
+                            <span class="avatar">
+                                <twig:ux:icon name="tabler:database-import" style="height: 2em;" />
+                            </span>
+                        </div>
+                        <div class="col">
+                            <div class="font-weight-medium">
+                                {{ importCount|number_format(0, ',', '.') }} {{ 'link.imports'|trans }}
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-6 col-lg-3">
+            <div class="card">
+                <div class="card-body">
+                    <div class="row align-items-center">
+                        <div class="col-auto">
+                            <span class="avatar">
+                                <twig:ux:icon name="tabler:user" style="height: 2em;" />
+                            </span>
+                        </div>
+                        <div class="col">
+                            <div class="font-weight-medium">
+                                {{ userCount }} {{ 'link.users'|trans }}
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-sm-6 col-lg-3">
+            <div class="card">
+                <div class="card-body">
+                    <div class="row align-items-center">
+                        <div class="col-auto">
+                            <span class="avatar">
+                                <twig:ux:icon name="tabler:building-hospital" style="height: 2em;" />
+                            </span>
+                        </div>
+                        <div class="col">
+                            <div class="font-weight-medium">
+                                {{ hospitalCount }} {{ 'link.hospitals'|trans }}
+                            </div>
+                            <div class="text-muted">
+                                {{ 'label.hospitals.participating'|trans }}:
+                                <strong>{{ participatingHospitalCount }}</strong>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div class="row row-cards mt-1"
+         data-controller="dashboard-charts"
+         data-dashboard-charts-allocations-value="{{ allocationChart|json_encode|e('html_attr') }}"
+         data-dashboard-charts-imports-value="{{ importChart|json_encode|e('html_attr') }}"
+    >
+        <div class="col-md-6">
+            <div class="card">
+                <div class="card-body">
+                    <h3 class="card-title mb-0">Allocations</h3>
+                    <div data-dashboard-charts-target="allocationsChart" class="chart-lg"></div>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-md-6">
+            <div class="card">
+                <div class="card-body">
+                    <h3 class="card-title mb-0">Imports</h3>
+                    <div data-dashboard-charts-target="importsChart" class="chart-lg"></div>
+                </div>
+            </div>
+        </div>
     </div>
 {% endblock %}
